Implements bulk generate feature for database server generators
to fix #47, however the larger the anonymous PL/SQL block gets the more
time the Oracle Database needs to parse it. The execution time is less
than a second for 7910 nodes but to parse the generated PL/SQL block it
takes more than 10 seconds. To improve the runtime and reduce the
resource consumption the nodes have to be passed as data and not as
code.